fio: fix str_to_ptr() crash at windows
I compiled a "fio_ceph_messenger. DLL" running on windows for FIO testing, but the original code is Linux, and the default size of long is 64 bits, but it is 32 bits on windows, resulting in a crash during pointer conversion. Now macro judgment is added, and it has been running well on my computer

Signed-off-by: wt.tao <i237731947@outlook.com>